Fixed leftover conflicts, moved nodemailer report to api router
