+++ app/index.js
@@ -21,14 +21,11 @@ const apiRouter = require('./routes/api');
 const passport = require('passport');
 const session = require('express-session');
 
-const db = require('better-sqlite3')('app.db');
 const SQLiteStore = require('connect-sqlite3')(session);
 
 const swaggerUI = require('swagger-ui-express');
 const yaml = require('yaml');
 const fs = require('fs');
-
-const nodemailer = require('nodemailer');
 
 // ================ SERVER SETUP ================
 app.set('view engine', 'ejs'); // Config express to use ejs as the "view engine" (See: https://expressjs.com/en/guide/using-template-engines.html)
@@ -67,21 +64,10 @@ app.use(function (req, res, next) {
     next();
 });
 
-// Configures email settings for reporting
-const transporter = nodemailer.createTransport({
-    host: process.env.EMAIL_HOST,
-    port: process.env.EMAIL_PORT,
-    auth: {
-        user: process.env.EMAIL_USER,
-        pass: process.env.EMAIL_PASS
-    }
-})
-
 // Configuring API docs via Swagger.
 const swaggerFile = fs.readFileSync('./swagger.yaml', 'utf-8');
 const swaggerDoc = yaml.parse(swaggerFile);
 app.use('/api-docs', swaggerUI.serve, swaggerUI.setup(swaggerDoc));
-
 
 // ================ ROUTERS ========================
 app.use('/', authRouter, clientRouter);
@@ -90,92 +76,6 @@ app.use('/api', apiRouter);
 // ================ JS AND CSS PATH SETUP ================
 app.use(express.static(path.join(__dirname, 'public/css')));
 app.use(express.static(path.join(__dirname, 'public/js')));
-
-// ================ APP ROUTES ================
-app.get('/', (_, res) => {
-    const pageContext = {
-        // Fake Fact data to mock landing page fact of the day, may be replaced with random fact?
-        factoid: {
-            id: 123,
-            content: 'A super cool 37 fact',
-            // note: 'Something extra about the fact',
-        },
-    };
-    res.render('pages/landing-page', pageContext);
-});
-
-// * This route is used for example EJS usage demonstration. Should be removed after we are familiar with EJS.
-app.get('/example', (_, res) => {
-    // Demonstrating passing an obj with values to a rendered template
-    const pageContext = { injectedVal: 'Superb!' };
-    res.render('pages/example', pageContext);
-});
-
-// This route is for the admin dashboard where the admin can approved, edit, and delete fact submissions.
-app.get('/admin', (req, res) => {
-    const testData = [
-        {
-            dateSubmitted: '05 / 04 / 2024',
-            user: 'abc0185',
-            fact: 'The number of pages in this book is a multiple of 37!',
-            note: 'Name of the book: "All things 37" by Greg Jones',
-            tags: 'media, books',
-        },
-        {
-            dateSubmitted: '05 / 06 / 2024',
-            user: 'mag3737',
-            fact: 'Another super cool 37 fact',
-            note: 'Found in Vancouver, BC',
-            tags: 'nature',
-        },
-    ];
-    const adminName = 'mag3737';
-    res.render('pages/admin-dashboard', { submissions: testData, adminName });
-});
-
-// This route is for the factoids listings page where users can view and search for factoids.
-app.get('/factoids', async (req, res) => {
-    const pageContext = {
-        // Fake Fact data to mock factoid, may be replaced with actual data from the database
-        factoid: {
-            id: 777,
-            content: 'A super cool 37 fact',
-            note: 'Something extra about the fact',
-        },
-    };
-    res.render('pages/factoid-listings', pageContext);
-});
-
-// This route is for the about/why 37? page.
-app.get('/about', async (req, res) => {
-    res.render('pages/about');
-});
-
-// This route displays contact information
-app.get('/contact', (req, res) => {
-    res.render('pages/contact');
-});
-
-app.post('/api/report', (req, res) => {
-    const reporter = res.locals.user.id;
-    const mailOptions = {
-        from: process.env.EMAIL_USER,
-        to: process.env.EMAIL_RECEIVER,
-        subject: "thirty-seven.org - Fact #" + req.body.fact.id +  " Has Been Reported",
-        text: "Reported by: " + reporter + "\nFact #" + req.body.fact.id + "\nFact: " + req.body.fact.content + "\n\n" + req.body.report.issue || "No issues!"
-    };
-    transporter.sendMail(mailOptions, (error, info) => {
-        if (error) {
-            console.error("Error sending email: ", error);
-        } else {
-            console.log("Email sent: ", info.response);
-        }
-    });
-    res.redirect("/factoids");
-})
-
-// ================ SERVER ROUTES ================
-// TODO: Add server REST route calls for making SQLite queries through prisma
 
 // Begin the server and listen on the configured port
 app.listen(PORT);
